Adds components Opponent, Card, Betting
App.js is currently filled with testing code for these three components,
and a folder of images of the 52 cards has also been added. I wasn't sure
what to do with the original src folder so I just replaced it with my
stuff. App.css has also been fillved with some styles for the
components.