Adds GameContainer and sample/testing state
Added GameContainer but the items currently in state are representations
of what we would need to take from the game state. As of now some of
those states aren't within the backend's game.js file. Still missing a
place for player to place down card and functions that would communicate
with rest of Oh Hell.